requirements.txt
requests
mysql-connector-python
httpx
fastapi[all]
